+++ backend/func.go
@@ -1,12 +1,12 @@
 package backend
 
 import (
-	"fmt"
+	"context"
 	"reflect"
 
 	"github.com/integration-system/isp-lib/v2/isp"
 	"github.com/integration-system/isp-lib/v2/streaming"
-	pkgerrors "github.com/pkg/errors"
+	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
@@ -15,8 +15,10 @@ import (
 type function struct {
 	dataParamType reflect.Type
 	mdParamType   reflect.Type
+	ctxParamNum   int
 	mdParamNum    int
 	dataParamNum  int
+	paramsCount   int
 	fun           reflect.Value
 	methodName    string
 }
@@ -40,25 +42,22 @@ func (f function) unmarshalAndValidateInputData(msg *isp.Message, ctx *ctx, vali
 	return nil, nil
 }
 
-func (f function) call(dataParam interface{}, md metadata.MD) (_ interface{}, err error) {
+func (f function) call(ctx context.Context, dataParam interface{}, md metadata.MD) (_ interface{}, err error) { // todo ctx context.Context
 	defer func() {
 		recovered := recover()
 		if recovered != nil {
-			err = pkgerrors.WithStack(fmt.Errorf("recovered panic from handler: %v", recovered))
+			err = errors.WithStack(errors.Errorf("recovered panic from handler: %v", recovered))
 		}
 	}()
-	var argCount int
-	if f.mdParamNum > f.dataParamNum {
-		argCount = f.mdParamNum + 1
-	} else {
-		argCount = f.dataParamNum + 1
-	}
-	args := make([]reflect.Value, argCount)
+	args := make([]reflect.Value, f.paramsCount)
 	if f.mdParamNum != -1 {
 		args[f.mdParamNum] = reflect.ValueOf(md).Convert(f.mdParamType)
 	}
 	if f.dataParamNum != -1 && dataParam != nil {
 		args[f.dataParamNum] = reflect.ValueOf(dataParam).Elem()
+	}
+	if f.ctxParamNum != -1 {
+		args[f.ctxParamNum] = reflect.ValueOf(ctx)
 	}
 
 	res := f.fun.Call(args)